Moved the testimonials to the protege-programs page.tsx
