+++ app/(pages)/programs/page.tsx
@@ -1,43 +1,8 @@
-import {Box, Stack, Typography} from '@mui/joy';
+import {Box, Stack} from '@mui/joy';
 import PageHeaderCard from '@/components/PageHeaderCard';
 import ProgramsCard from '@/components/ProgramsCard';
 import {IconTextProps} from '@/components/LabelledIcon';
 import HeadingBodyText from '@/components/HeadingBodyText';
-import HorizontalLine from '@/components/HorizontalLine';
-import Quote from '@/components/Quote';
-import { protegeData } from '@/app/data';
-
-function Testimonials() {
-  return (
-    <Stack
-      py={5}
-      px={5}
-      mx='auto'
-      textAlign='center'
-      justifyContent='center'
-      maxWidth={1250}
-      width='100%'
-      alignItems='center'
-    >
-      <Typography level='subtitle-lg' mb={5}>
-        Hear from our previous mentors and mentees
-      </Typography>
-      <Stack spacing={5} direction='column' width='100%'>
-        <Box display='flex' justifyContent='center' width='100%'>
-          <HorizontalLine color='#93AED3' height={2} width={1200} />
-        </Box>
-        {protegeData.map(({ quote, by }, idx) => (
-          <Stack key={idx} spacing={5} maxWidth={1250} width='100%' alignItems='center'>
-            <Quote quote={quote} color='dark-blue' author={by} />
-            <Box display='flex' justifyContent='center' width='100%'>
-              <HorizontalLine color='#93AED3' height={2} width={1200} />
-            </Box>
-          </Stack>
-        ))}
-      </Stack>
-    </Stack>
-  );
-}
 
 export default function Program() {
   const bodyText =
@@ -93,7 +58,6 @@ export default function Program() {
                       description={impDescription} icon1={impIcon1} icon2={impIcon2} icon3={impIcon3}
                       icon4={impIcon4} link='/industry-mentoring-program'/>
       </Stack>
-      <Testimonials />
     </Box>
   );
 }
