+++ app/(pages)/programs/protege-program/page.tsx
@@ -1,8 +1,34 @@
-import {Box, Stack, AspectRatio} from '@mui/joy';
+import {Box, Stack, AspectRatio, Typography} from '@mui/joy';
 import PageHeaderCard from '@/components/PageHeaderCard';
 import Image from 'next/image';
 import HeadingBodyText from '@/components/HeadingBodyText';
 import ProgramsTitleDescription from '@/components/ProgramsTitleDescription';
+import HorizontalLine from '@/components/HorizontalLine';
+import Quote from '@/components/Quote';
+import { protegeData } from '@/app/data';
+
+function Testimonials() {
+  return (
+    <Stack py={5} px={5} mx='auto' textAlign='center' justifyContent='center' maxWidth={1250} width='100%' alignItems='center'>
+      <Typography level='subtitle-lg' mb={5}>
+        Hear from our previous mentors and mentees
+      </Typography>
+      <Stack spacing={5} direction='column' width='100%'>
+        <Box display='flex' justifyContent='center' width='100%'>
+          <HorizontalLine color='#93AED3' height={2} width={1200} />
+        </Box>
+        {protegeData.map(({ quote, by }, idx) => (
+          <Stack key={idx} spacing={5} maxWidth={1250} width='100%' alignItems='center'>
+            <Quote quote={quote} color='dark-blue' author={by} />
+            <Box display='flex' justifyContent='center' width='100%'>
+              <HorizontalLine color='#93AED3' height={2} width={1200} />
+            </Box>
+          </Stack>
+        ))}
+      </Stack>
+    </Stack>
+  );
+}
 
 export default function ProtegeProgram() {
 
@@ -51,6 +77,8 @@ export default function ProtegeProgram() {
 
         </Stack>
       </Box>
+
+      <Testimonials />
     </Box>
   );
 }
